+++ games/poker/scenes/PlayScene.ts
@@ -1,10 +1,10 @@
 /* eslint-disable class-methods-use-this */
-import { Result } from '@/games/common/types/game';
 import STYLE from '@/games/common/constants/style';
+import { Result } from '@/games/common/types/game';
 import Button from '../../common/Factories/button';
-import Card from '../../common/Factories/cardImage';
-import Deck from '../../common/Factories/deckImage';
-import Table from '../../common/Factories/tableScene';
+import Card from '../../common/Factories/card';
+import Deck from '../../common/Factories/deck';
+import Table from '../../common/Factories/table';
 import GAME from '../../common/constants/game';
 import GameResult from '../constants/gameResult';
 import GameStatus from '../constants/gameStatus';
